refactor: playground 관련 코드 이동 및 PlaygroundHeaderCreator 추 (#324)
