+++ src/main/java/org/sopt/app/application/playground/PlaygroundUserRecommenderRedisImpl.java
@@ -1,7 +1,8 @@
-package org.sopt.app.application.auth;
+package org.sopt.app.application.playground;
+
+import static org.sopt.app.application.playground.PlaygroundHeaderCreator.createAuthorizationHeaderByInternalPlaygroundToken;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.sopt.app.application.auth.dto.RecommendFriendRequest;
@@ -9,16 +10,12 @@ import org.sopt.app.application.auth.dto.RecommendedFriendInfo.PlaygroundUserIds
 import org.sopt.app.domain.entity.RecommendedUsers;
 import org.sopt.app.interfaces.external.PlaygroundClient;
 import org.sopt.app.interfaces.postgres.RecommendedUserIdsRepository;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class PlaygroundUserRecommenderRedisImpl implements PlaygroundUserRecommender {
 
-    @Value("${makers.playground.access-token}")
-    private String playgroundToken;
     private final PlaygroundClient playgroundClient;
     private final RecommendedUserIdsRepository recommendedUserIdsRepository;
 
@@ -31,13 +28,10 @@ public class PlaygroundUserRecommenderRedisImpl implements PlaygroundUserRecomme
         }
 
         PlaygroundUserIds playgroundUserIds =
-                playgroundClient.getPlaygroundUserIdsForSameRecommendType(this.getAuthHeader(), request);
-        return recommendedUserIdsRepository.save(new RecommendedUsers(request, playgroundUserIds.userIds()))
-                .getUserIds();
-    }
-
-    private Map<String, String> getAuthHeader(){
-      return Map.of(HttpHeaders.AUTHORIZATION, playgroundToken);
+                playgroundClient.getPlaygroundUserIdsForSameRecommendType(
+                        createAuthorizationHeaderByInternalPlaygroundToken(), request);
+        recommendedUserIdsRepository.save(new RecommendedUsers(request, playgroundUserIds.userIds()));
+        return playgroundUserIds.userIds();
     }
 
 }
