+++ src/main/java/org/sopt/app/presentation/user/UserOriginalController.java
@@ -8,12 +8,11 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
-import org.sopt.app.application.auth.PlaygroundAuthService;
+import org.sopt.app.application.playground.PlaygroundAuthService;
 import org.sopt.app.application.auth.dto.PlaygroundPostInfo.PlaygroundPost;
 import org.sopt.app.domain.entity.User;
 import org.sopt.app.facade.UserFacade;
 import org.sopt.app.presentation.user.UserResponse.AppService;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
